// src/common/enum/env.enum.ts
export enum ENVEnum {
  NODE_ENV = 'NODE_ENV',
  DATABASE_URL = 'DATABASE_URL',
  PORT = 'PORT',
  BASE_URL = 'BASE_URL',
  JWT_ACCESS_SECRET = 'JWT_ACCESS_SECRET',
  JWT_ACCESS_EXPIRES_IN = 'JWT_ACCESS_EXPIRES_IN',
  JWT_REFRESH_EXPIRES_IN = 'JWT_REFRESH_EXPIRES_IN',
  JWT_REFRESH_SECRET = 'JWT_REFRESH_SECRET',

  FRONTEND_URL = 'FRONTEND_URL',

  GOOGLE_CLIENT_ID = 'GOOGLE_CLIENT_ID',
  GOOGLE_CLIENT_SECRET = 'GOOGLE_CLIENT_SECRET',
  GOOGLE_CALLBACK_URL = 'GOOGLE_CALLBACK_URL',

  MAIL_HOST = 'MAIL_HOST',
  MAIL_PORT = 'MAIL_PORT',
  MAIL_USER = 'MAIL_USER',
  MAIL_PASS = 'MAIL_PASS',
  MAIL_FROM = 'MAIL_FROM',

  CLOUDINARY_CLOUD_NAME = 'CLOUDINARY_CLOUD_NAME',
  CLOUDINARY_API_KEY = 'CLOUDINARY_API_KEY',
  CLOUDINARY_API_SECRET = 'CLOUDINARY_API_SECRET',

  REDIS_HOST = 'REDIS_HOST',
  REDIS_PORT = 'REDIS_PORT',
  REDIS_USERNAME = 'REDIS_USERNAME',
  REDIS_PASSWORD = 'REDIS_PASSWORD',

  STRIPE_SECRET_KEY = 'STRIPE_SECRET_KEY',
  STRIPE_WEBHOOK_SECRET = 'STRIPE_WEBHOOK_SECRET',

  AGORA_APP_ID = 'AGORA_APP_ID',
  AGORA_APP_CERTIFICATE = 'AGORA_APP_CERTIFICATE',
  AGORA_TOKEN_EXPIRE_SECONDS = 'AGORA_TOKEN_EXPIRE_SECONDS',

  GOOGLE_MAPS_API_KEY = 'GOOGLE_MAPS_API_KEY',

  FIREBASE_ENABLED = 'FIREBASE_ENABLED',
  FIREBASE_PROJECT_ID = 'FIREBASE_PROJECT_ID',
  FIREBASE_CLIENT_EMAIL = 'FIREBASE_CLIENT_EMAIL',
  FIREBASE_PRIVATE_KEY = 'FIREBASE_PRIVATE_KEY',

  SUPER_ADMIN_EMAIL = 'SUPER_ADMIN_EMAIL',
  SUPER_ADMIN_PASS = 'SUPER_ADMIN_PASS',

  AWS_REGION = 'AWS_REGION',
  AWS_ACCESS_KEY_ID = 'AWS_ACCESS_KEY_ID',
  AWS_SECRET_ACCESS_KEY = 'AWS_SECRET_ACCESS_KEY',
  AWS_S3_BUCKET_NAME = 'AWS_S3_BUCKET_NAME',
}